// src/infra/api/model/enderecoModel.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Minha API
 * Documentação oficial da API
 * OpenAPI spec version: 1.0.0
 */

export interface EnderecoModel {
  id?: number;
  estabelecimentoId: number;
  /** @maxLength 255 */
  rua: string;
  /** @maxLength 20 */
  numero: string;
  /** @maxLength 255 */
  complemento: string;
  /** @maxLength 150 */
  bairro: string;
  /** @maxLength 150 */
  cidade: string;
  /** @maxLength 100 */
  estado: string;
  /** @maxLength 20 */
  cep: string;
  latitude: string;
  longitude: string;
}