// src/infra/api/model/servicoModel.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Minha API
 * Documentação oficial da API
 * OpenAPI spec version: 1.0.0
 */

export interface ServicoModel {
  id: number;
  /** @nullable */
  titulo: string | null;
  /** @nullable */
  descricao: string | null;
  /** @nullable */
  valor: string | null;
  /** @nullable */
  estabelecimentoId: number | null;
}